move: com.skyhorsemanpower group으로 변경
