+++ src/main/java/com/skyhorsemanpower/payment/PaymentApplication.java
@@ -1,0 +1,17 @@
+package com.skyhorsemanpower.payment;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.cloud.context.config.annotation.RefreshScope;
+
+@SpringBootApplication
+@RefreshScope
+@EnableDiscoveryClient
+public class PaymentApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(PaymentApplication.class, args);
+	}
+
+}
