+++ src/main/java/com/skyhorsemanpower/payment/common/CommonResponse.java
@@ -1,0 +1,10 @@
+package com.skyhorsemanpower.payment.common;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class CommonResponse<T> {
+
+}
