+++ src/main/java/com/skyhorsemanpower/payment/common/MemberPaymentStatus.java
@@ -1,0 +1,12 @@
+package com.skyhorsemanpower.payment.common;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum MemberPaymentStatus {
+	MEMBER_PAYMENT_AGREE,
+	MEMBER_PAYMENT_DISAGREE,
+	MEMBER_PAYMENT_READY
+}
