+++ src/main/java/com/skyhorsemanpower/payment/common/PaymentStatus.java
@@ -1,0 +1,15 @@
+package com.skyhorsemanpower.payment.common;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum PaymentStatus {
+
+	PAYMENT_COMPLETE,
+	PAYMENT_FAIL,
+	PAYMENT_CANCEL,
+	PAYMENT_PENDING,
+	PAYMENT_READY
+}
