+++ src/main/java/com/skyhorsemanpower/payment/common/SuccessResponse.java
@@ -1,0 +1,13 @@
+package com.skyhorsemanpower.payment.common;
+
+import lombok.Builder;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+
+public class SuccessResponse<T> extends ResponseEntity<T> {
+
+	@Builder
+	public SuccessResponse(T body) {
+		super(body, HttpStatus.OK);
+	}
+}
