+++ src/main/java/com/skyhorsemanpower/payment/common/exception/CustomException.java
@@ -1,0 +1,16 @@
+package com.skyhorsemanpower.payment.common.exception;
+
+public class CustomException extends RuntimeException {
+
+    private final ResponseStatus responseStatus;
+
+    public CustomException(ResponseStatus responseStatus) {
+        super(responseStatus.getMessage());
+        this.responseStatus = responseStatus;
+    }
+
+    public ResponseStatus getResponseStatus() {
+        return responseStatus;
+    }
+
+}
