+++ src/main/java/com/skyhorsemanpower/payment/common/exception/GlobalExceptionHandler.java
@@ -1,0 +1,15 @@
+package com.skyhorsemanpower.payment.common.exception;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+@ControllerAdvice
+public class GlobalExceptionHandler {
+
+    @ExceptionHandler(CustomException.class)
+    public ResponseEntity<String> handleCustomException(CustomException e) {
+        ResponseStatus responseStatus = e.getResponseStatus();
+        return new ExceptionResponse(responseStatus);
+    }
+}
