+++ src/main/java/com/skyhorsemanpower/payment/payment/dto/PaymentAgreeRequestDto.java
@@ -1,0 +1,24 @@
+package com.skyhorsemanpower.payment.payment.dto;
+
+import com.skyhorsemanpower.payment.payment.vo.PaymentAgreeRequestVo;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class PaymentAgreeRequestDto {
+
+	private String paymentUuid;
+
+	public static PaymentAgreeRequestDto voToDto(PaymentAgreeRequestVo paymentAgreeRequestVo) {
+		return PaymentAgreeRequestDto.builder()
+			.paymentUuid(paymentAgreeRequestVo.getPaymentUuid())
+			.build();
+	}
+}
