+++ src/main/java/com/skyhorsemanpower/payment/payment/dto/PaymentDetailRequestDto.java
@@ -1,0 +1,24 @@
+package com.skyhorsemanpower.payment.payment.dto;
+
+import com.skyhorsemanpower.payment.payment.vo.PaymentDetailRequestVo;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class PaymentDetailRequestDto {
+
+	private String paymentUuid;
+
+	public static PaymentDetailRequestDto voToDto(PaymentDetailRequestVo paymentDetailRequestVo) {
+		return PaymentDetailRequestDto.builder()
+			.paymentUuid(paymentDetailRequestVo.getPaymentUuid())
+			.build();
+	}
+}
