+++ src/main/java/com/skyhorsemanpower/payment/payment/vo/PaymentAgreeRequestVo.java
@@ -1,0 +1,9 @@
+package com.skyhorsemanpower.payment.payment.vo;
+
+import lombok.Getter;
+
+@Getter
+public class PaymentAgreeRequestVo {
+
+	private String paymentUuid;
+}
