+++ src/main/java/com/skyhorsemanpower/payment/payment/vo/PaymentListResponseVo.java
@@ -1,0 +1,24 @@
+package com.skyhorsemanpower.payment.payment.vo;
+
+import com.skyhorsemanpower.payment.common.PaymentStatus;
+import java.time.LocalDateTime;
+import lombok.Getter;
+
+@Getter
+public class PaymentListResponseVo {
+
+	private String paymentUuid;
+	private String auctionUuid;
+	private int price;
+	private PaymentStatus paymentStatus;
+	private LocalDateTime paymentAt;
+
+	public PaymentListResponseVo(String paymentUuid, String auctionUuid, int price,
+		PaymentStatus paymentStatus, LocalDateTime paymentAt) {
+		this.paymentUuid = paymentUuid;
+		this.auctionUuid = auctionUuid;
+		this.price = price;
+		this.paymentStatus = paymentStatus;
+		this.paymentAt = paymentAt;
+	}
+}
