--- src/main/java/com/sparos4th2/payment/PaymentApplication.java
@@ -1,17 +1,0 @@
-package com.sparos4th2.payment;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.cloud.context.config.annotation.RefreshScope;
-
-@SpringBootApplication
-@RefreshScope
-@EnableDiscoveryClient
-public class PaymentApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PaymentApplication.class, args);
-	}
-
-}
