--- src/main/java/com/sparos4th2/payment/common/CommonResponse.java
@@ -1,10 +1,0 @@
-package com.sparos4th2.payment.common;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class CommonResponse<T> {
-
-}
