--- src/main/java/com/sparos4th2/payment/common/SuccessResponse.java
@@ -1,13 +1,0 @@
-package com.sparos4th2.payment.common;
-
-import lombok.Builder;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-
-public class SuccessResponse<T> extends ResponseEntity<T> {
-
-	@Builder
-	public SuccessResponse(T body) {
-		super(body, HttpStatus.OK);
-	}
-}
