--- src/main/java/com/sparos4th2/payment/common/exception/CustomException.java
@@ -1,16 +1,0 @@
-package com.sparos4th2.payment.common.exception;
-
-public class CustomException extends RuntimeException {
-
-    private final ResponseStatus responseStatus;
-
-    public CustomException(ResponseStatus responseStatus) {
-        super(responseStatus.getMessage());
-        this.responseStatus = responseStatus;
-    }
-
-    public ResponseStatus getResponseStatus() {
-        return responseStatus;
-    }
-
-}
