--- src/main/java/com/sparos4th2/payment/common/exception/ExceptionResponse.java
@@ -1,12 +1,0 @@
-package com.sparos4th2.payment.common.exception;
-
-import lombok.Builder;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-
-public class ExceptionResponse extends ResponseEntity<String> {
-    @Builder
-    public ExceptionResponse(ResponseStatus responseStatus) {
-        super(responseStatus.getMessage(), HttpStatus.valueOf(responseStatus.getCode()));
-    }
-}
