--- src/main/java/com/sparos4th2/payment/common/exception/GlobalExceptionHandler.java
@@ -1,15 +1,0 @@
-package com.sparos4th2.payment.common.exception;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-@ControllerAdvice
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(CustomException.class)
-    public ResponseEntity<String> handleCustomException(CustomException e) {
-        ResponseStatus responseStatus = e.getResponseStatus();
-        return new ExceptionResponse(responseStatus);
-    }
-}
