--- src/main/java/com/sparos4th2/payment/common/swagger/SwaggerConfig.java
@@ -1,46 +1,0 @@
-package com.sparos4th2.payment.common.swagger;
-
-import io.swagger.v3.oas.annotations.OpenAPIDefinition;
-import io.swagger.v3.oas.annotations.info.Info;
-import io.swagger.v3.oas.annotations.servers.Server;
-import io.swagger.v3.oas.models.Components;
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.security.SecurityRequirement;
-import io.swagger.v3.oas.models.security.SecurityScheme;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@OpenAPIDefinition(
-    info = @Info(
-        title = "Payment Service API",
-        version = "v1"
-    ),
-    servers = @Server(
-        url = "http://localhost:8080",
-        description = "Local Server"
-    )
-)
-
-@Configuration
-public class SwaggerConfig {
-
-    private static final String BEARER_TOKEN_PREFIX = "Bearer";
-
-    @Bean
-    public OpenAPI openAPI() {
-
-        String securityJwtName = "JWT";
-        SecurityRequirement securityRequirement = new SecurityRequirement().addList(
-            securityJwtName);
-        Components components = new Components()
-            .addSecuritySchemes(securityJwtName, new SecurityScheme()
-                .name(securityJwtName)
-                .type(SecurityScheme.Type.HTTP)
-                .scheme(BEARER_TOKEN_PREFIX)
-                .bearerFormat(securityJwtName));
-
-        return new OpenAPI()
-            .addSecurityItem(securityRequirement)
-            .components(components);
-    }
-}
