--- src/main/java/com/sparos4th2/payment/payment/vo/PaymentAgreeRequestVo.java
@@ -1,9 +1,0 @@
-package com.sparos4th2.payment.payment.vo;
-
-import lombok.Getter;
-
-@Getter
-public class PaymentAgreeRequestVo {
-
-	private String paymentUuid;
-}
