+++ src/test/java/com/skyhorsemanpower/payment/PaymentApplicationTests.java
@@ -1,0 +1,13 @@
+package com.skyhorsemanpower.payment;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PaymentApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
