--- src/test/java/com/sparos4th2/payment/PaymentApplicationTests.java
@@ -1,13 +1,0 @@
-package com.sparos4th2.payment;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-//@SpringBootTest
-class PaymentApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
